datetimecardのcssの削除とフォ削除のためのファイルパスの変更
