--- frontend/src/components/dates/dateTimeCard.css
@@ -1,70 +1,0 @@
-/* dateTimeCard.css */
-
-.date-card {
-  background: #fff;
-  border-radius: 12px;
-  padding: 16px;
-  margin-bottom: 16px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.06);
-}
-
-.section {
-  margin-bottom: 12px;
-}
-
-.section-title {
-  font-size: 13px;
-  color: #555;
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  margin-bottom: 8px;
-}
-
-.time-row {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin-bottom: 6px;
-}
-
-.label {
-  width: 48px;
-  font-size: 14px;
-  color: #333;
-}
-
-.time-button {
-  padding: 6px 12px;
-  border-radius: 8px;
-  border: 1px solid #ccc;
-  background: #f9fafb;
-  cursor: pointer;
-  font-size: 14px;
-}
-
-.time-text {
-  font-size: 14px;
-  color: #222;
-}
-
-.date-list {
-  padding-left: 18px;
-  margin: 0;
-}
-
-.date-list li {
-  font-size: 14px;
-  margin-bottom: 4px;
-}
-
-/* dateTimeCard.css */
-.date-item {
-  font-size: 14px;
-  margin-bottom: 4px;
-}
-
-.date-item--past {
-  color: #999;
-  opacity: 0.6;
-}
